+++ singer/__init__.py
@@ -61,8 +61,10 @@ from singer.catalog import (
 )
 from singer.schema import Schema
 
-from singer.bookmarks import (
-    write_bookmark,
+from singer.state import (
+    set_bookmark,
+    # for backwards compatibility, use set_bookmark instead
+    set_bookmark as write_bookmark,
     get_bookmark,
     clear_bookmark,
     reset_stream,
@@ -71,6 +73,9 @@ from singer.bookmarks import (
     get_offset,
     set_currently_syncing,
     get_currently_syncing,
+    set_version,
+    clear_version,
+    get_version,
 )
 
 from singer.exceptions import (