+++ tests/test_state.py
@@ -1,7 +1,7 @@
 import unittest
 from singer import state as st
 
-class TestGetBookmark(unittest.TestCase):
+class TestBookmark(unittest.TestCase):
     def test_empty_state(self):
         empty_state = {}
 
@@ -69,8 +69,24 @@ class TestGetBookmark(unittest.TestCase):
         self.assertEqual(st.get_bookmark(non_empty_state, stream_id_1, bookmark_key_1, 'default_value'),
                          bookmark_val_1)
 
+    def test_set_bookmark(self):
+        stream_id_1 = 'customers'
+        bookmark_key_1 = 'datetime'
+        bookmark_val_1 = 123456789
 
-class TestGetOffset(unittest.TestCase):
+        result = st.set_bookmark({'bookmarks': {stream_id_1: {bookmark_key_1: 'old-value'}}}, stream_id_1, bookmark_key_1, bookmark_val_1)
+        self.assertEqual(result, {'bookmarks': {stream_id_1: {bookmark_key_1: bookmark_val_1}}})
+
+    def test_clear_bookmark(self):
+        stream_id_1 = 'customers'
+        bookmark_key_1 = 'datetime'
+        bookmark_val_1 = 123456789
+
+        result = st.clear_bookmark({'bookmarks': {stream_id_1: {bookmark_key_1: bookmark_val_1}}}, stream_id_1, bookmark_key_1)
+        self.assertEqual(result, {'bookmarks': {stream_id_1: {}}})
+
+
+class TestOffset(unittest.TestCase):
     def test_empty_state(self):
         empty_state = {}
 
@@ -129,8 +145,24 @@ class TestGetOffset(unittest.TestCase):
         self.assertEqual(st.get_offset(non_empty_state, stream_id_1, 'default_value'),
                          offset_val)
 
+    def test_set_offset(self):
+        stream_id_1 = 'customers'
+        offset_key_1 = 'datetime'
+        offset_val_1 = 123456789
 
-class TestGetCurrentlySyncing(unittest.TestCase):
+        result = st.set_offset({'bookmarks': {stream_id_1: {'offset': {offset_key_1: 'old-value'}}}}, stream_id_1, offset_key_1, offset_val_1)
+        self.assertEqual(result, {'bookmarks': {stream_id_1: {'offset': {offset_key_1: offset_val_1}}}})
+
+    def test_clear_offset(self):
+        stream_id_1 = 'customers'
+        offset_key_1 = 'datetime'
+        offset_val_1 = 123456789
+
+        result = st.clear_offset({'bookmarks': {stream_id_1: {'offset': {offset_key_1: offset_val_1}}}}, stream_id_1)
+        self.assertEqual(result, {'bookmarks': {stream_id_1: {}}})
+
+
+class TestCurrentlySyncing(unittest.TestCase):
     def test_empty_state(self):
         empty_state = {}
 
@@ -164,6 +196,10 @@ class TestGetCurrentlySyncing(unittest.TestCase):
         # Case with a given default
         self.assertEqual(st.get_currently_syncing(non_empty_state, 'default_value'),
                          stream_id_1)
+
+    def test_set_currently_syncing(self):
+        result = st.set_currently_syncing({'currently_syncing': 'foo'}, 'bar')
+        self.assertEqual(result, {'currently_syncing': 'bar'})
 
 
 class TestActivateVersion(unittest.TestCase):
@@ -233,3 +269,19 @@ class TestActivateVersion(unittest.TestCase):
         # Good stream, good key
         self.assertEqual(st.get_version(non_empty_state, stream_id_1, version_key_1, 'default_value'),
                          version_val_1)
+
+    def test_set_version(self):
+        stream_id_1 = 'customers'
+        version_key_1 = 'datetime'
+        version_val_1 = 123456789
+
+        result = st.set_version({'activate_versions': {stream_id_1: {version_key_1: 'old-value'}}}, stream_id_1, version_key_1, version_val_1)
+        self.assertEqual(result, {'activate_versions': {stream_id_1: {version_key_1: version_val_1}}})
+
+    def test_clear_version(self):
+        stream_id_1 = 'customers'
+        version_key_1 = 'datetime'
+        version_val_1 = 123456789
+
+        result = st.clear_version({'activate_versions': {stream_id_1: {version_key_1: version_val_1}}}, stream_id_1, version_key_1)
+        self.assertEqual(result, {'activate_versions': {stream_id_1: {}}})